# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OptiMeal API",
    "description": "\n**ADR-008 이후의 계약**: 본 API의 스케일링은 정확 예측기가 아니다.\n\n* cold-start 는 **선형(b=1)** — 실측 OOS에서 최선·무편향이었기 때문이다(과소예측 위험 없음).\n* 정확도는 영양사 보정을 `/api/calibration/observations` 로 **누적**하면서 개선된다.\n* CBR 참고 이력은 표시 전용이며 **자동 적용하지 않는다**(`auto_apply=false`).\n* 모든 스케일링 응답은 재료별 `method`·`n_obs`·`confidence` 를 동반한다(감사 가능성).\n\n---\n### 프론트엔드 개발자용 연동 안내\n* **Base URL(개발)**: `http://localhost:8000`\n* **CORS**: `localhost:5173`(Vite)·`localhost:3000`(CRA) 허용. 운영은 `OPTIMEAL_CORS_ORIGINS`(콤마 구분)로 지정.\n* **인증**: 현재 없음(학내 프로젝트). 영양사 식별은 요청의 `site_id`로 구분.\n* **선택적 의존성**: `/api/nutrition/*`·`/api/menu/*`는 PostgreSQL·모듈3 필요. 미구성 시 **503 + reason·hint** 반환(500 아님) → 프론트는 503을 \"인프라 준비중\"으로 처리.\n* **핵심 흐름**: ① 레시피 목록(`/api/scaling/recipes`) → ② 스케일링(`/api/scaling/predict`, `site_id`를 주면 보정 반영) → ③ 영양사 보정 누적(`/api/calibration/observations`). 스케일링 응답의 재료별 `method`·`confidence`를 UI에 표시할 것.\n* **알려진 미구현**: 식단가(원가)·알레르기 배제는 데이터(`recipe_ingredient_map`·`constraints`) 적재 후 활성화 예정.\n",
    "contact": {
      "name": "권성민 (Back)"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "개발 서버"
    }
  ],
  "paths": {
...
    "/api/scaling/recipes": {
      "get": {
        "tags": [
          "scaling"
        ],
        "summary": "레시피 목록 (키↔정수 id 매핑 포함)",
        "description": "스케일링 대상 레시피 목록. 캘리브레이션 API가 쓰는 정수 recipe_id를 함께 준다.",
        "operationId": "list_recipes_api_scaling_recipes_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "레시피 키 부분일치",
              "default": "",
              "title": "Q"
            },
            "description": "레시피 키 부분일치"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Recipes Api Scaling Recipes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scaling/recipes/{recipe_key}": {
      "get": {
        "tags": [
          "scaling"
        ],
        "summary": "레시피 재료 구성",
        "description": "레시피 1건의 재료 구성(1인분 기준)을 반환한다.\n\nRaises:\n    HTTPException(404): 미등록 레시피 키.",
        "operationId": "get_recipe_api_scaling_recipes__recipe_key__get",
        "parameters": [
          {
            "name": "recipe_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Recipe Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Recipe Api Scaling Recipes  Recipe Key  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scaling/predict": {
      "post": {
        "tags": [
          "scaling"
        ],
        "summary": "레시피 스케일링 (cold-start 선형 + 캘리브레이션 오버레이)",
        "description": "소규모 레시피를 목표 인원수로 변환한다.\n\nsite_id를 주지 않으면 전 재료가 cold-start 선형(`base×N`)이다. 주면 그 업장에 누적된\n보정이 있는 재료만 `calibrated`로 바뀐다 — 어떤 재료가 무슨 근거로 계산됐는지\n응답에서 재료 단위로 구분된다(감사 가능성).\n\nArgs:\n    payload: recipe_key·n_target·site_id·CBR 옵션.\n\nReturns:\n    재료별 스케일링 결과 + 근거 + ADR-008 고지.\n\nRaises:\n    HTTPException(404): 미등록 레시피 키.",
        "operationId": "predict_api_scaling_predict_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScalingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScalingResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scaling/coefficients": {
      "get": {
        "tags": [
          "scaling"
        ],
        "summary": "스케일링 계수 조회 (학술 비교군 — 운영 미적용)",
        "description": "`scaling_coefficients.csv`(331행, 수정 금지) 조회.\n\nADR-008 이후 이 계수는 운영 스케일링에 **적용되지 않는다**. 논문 비교군·이력 추적용.",
        "operationId": "get_coefficients_api_scaling_coefficients_get",
        "parameters": [
          {
            "name": "group_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "3대분류 필터",
              "default": "",
              "title": "Group Type"
            },
            "description": "3대분류 필터"
          },
          {
            "name": "ingredient_category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "재료 카테고리 필터",
              "default": "",
              "title": "Ingredient Category"
            },
            "description": "재료 카테고리 필터"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Coefficients Api Scaling Coefficients Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/nutrition/search": {
      "get": {
        "tags": [
          "nutrition"
        ],
        "summary": "영양성분 검색 (메뉴명)",
        "description": "메뉴명 부분일치로 영양성분을 검색한다.\n\nRaises:\n    HTTPException(503): 영양성분 DB 미구성.",
        "operationId": "search_api_nutrition_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "메뉴명 검색어",
              "title": "Q"
            },
            "description": "메뉴명 검색어"
          },
          {
            "name": "menu_category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "주식/국/찌개/주찬/부찬/김치 등",
              "default": "",
              "title": "Menu Category"
            },
            "description": "주식/국/찌개/주찬/부찬/김치 등"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Search Api Nutrition Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/nutrition/{nutrition_id}": {
      "get": {
        "tags": [
          "nutrition"
        ],
        "summary": "영양성분 상세 조회",
        "description": "영양성분 1건을 조회한다.\n\nRaises:\n    HTTPException(404): 미존재 id.\n    HTTPException(503): 영양성분 DB 미구성.",
        "operationId": "detail_api_nutrition__nutrition_id__get",
        "parameters": [
          {
            "name": "nutrition_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Nutrition Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Detail Api Nutrition  Nutrition Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CbrReference": {
        "properties": {
          "site_id": {
            "type": "integer",
            "title": "Site Id"
          },
          "recipe_id": {
            "type": "integer",
            "title": "Recipe Id"
          },
          "jaccard": {
            "type": "number",
            "title": "Jaccard",
            "description": "재료구성 Jaccard 유사도"
          },
          "est_ratio": {
            "type": "number",
            "title": "Est Ratio"
          },
          "n_obs": {
            "type": "integer",
            "title": "N Obs"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "recipe_id",
          "jaccard",
          "est_ratio",
          "n_obs"
        ],
        "title": "CbrReference",
        "description": "CBR 참고 이력 1건(표시 전용)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScaledIngredient": {
        "properties": {
          "ingredient_id": {
            "type": "integer",
            "title": "Ingredient Id"
          },
          "ingredient_name": {
            "type": "string",
            "title": "Ingredient Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "base_amount_g": {
            "type": "number",
            "title": "Base Amount G"
          },
          "scaled_g": {
            "type": "number",
            "title": "Scaled G"
          },
          "ratio": {
            "type": "number",
            "title": "Ratio"
          },
          "method": {
            "type": "string",
            "enum": [
              "cold_start_linear",
              "calibrated"
            ],
            "title": "Method"
          },
          "n_obs": {
            "type": "integer",
            "title": "N Obs"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "cold_start",
              "low",
              "high"
            ],
            "title": "Confidence"
          },
          "cbr_references": {
            "items": {
              "$ref": "#/components/schemas/CbrReference"
            },
            "type": "array",
            "title": "Cbr References"
          }
        },
        "type": "object",
        "required": [
          "ingredient_id",
          "ingredient_name",
          "role",
          "base_amount_g",
          "scaled_g",
          "ratio",
          "method",
          "n_obs",
          "confidence"
        ],
        "title": "ScaledIngredient",
        "description": "재료별 스케일링 결과 — 근거·신뢰도를 항상 동반한다."
      },
      "ScalingRequest": {
        "properties": {
          "recipe_key": {
            "type": "string",
            "title": "Recipe Key",
            "description": "df_B.csv small_recipe_id (예: A1034)"
          },
          "n_target": {
            "type": "integer",
            "maximum": 5000.0,
            "minimum": 1.0,
            "title": "N Target",
            "description": "목표 인원수 N"
          },
          "site_id": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id",
            "description": "업장 id. 주면 그 업장 보정을 적용"
          },
          "include_cbr": {
            "type": "boolean",
            "title": "Include Cbr",
            "description": "재료별 CBR 참고 이력 동봉 여부",
            "default": false
          },
          "cbr_top_k": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Cbr Top K",
            "default": 3
          }
        },
        "type": "object",
        "required": [
          "recipe_key",
          "n_target"
        ],
        "title": "ScalingRequest",
        "description": "레시피 스케일링 요청.",
        "example": {
          "cbr_top_k": 3,
          "include_cbr": true,
          "n_target": 100,
          "recipe_key": "A1034",
          "site_id": 5
        }
      },
      "ScalingResponse": {
        "properties": {
          "recipe_key": {
            "type": "string",
            "title": "Recipe Key"
          },
          "recipe_id": {
            "type": "integer",
            "title": "Recipe Id"
          },
          "group_type": {
            "type": "string",
            "title": "Group Type"
          },
          "cooking_method": {
            "type": "string",
            "title": "Cooking Method"
          },
          "n_target": {
            "type": "integer",
            "title": "N Target"
          },
          "site_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Site Id"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/ScaledIngredient"
            },
            "type": "array",
            "title": "Ingredients"
          },
          "calibrated_count": {
            "type": "integer",
            "title": "Calibrated Count",
            "description": "캘리브레이션이 적용된 재료 수"
          },
          "cold_start_count": {
            "type": "integer",
            "title": "Cold Start Count",
            "description": "cold-start 선형으로 계산된 재료 수"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer"
          }
        },
        "type": "object",
        "required": [
          "recipe_key",
          "recipe_id",
          "group_type",
          "cooking_method",
          "n_target",
          "site_id",
          "ingredients",
          "calibrated_count",
          "cold_start_count",
          "disclaimer"
        ],
        "title": "ScalingResponse",
        "description": "스케일링 응답."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}